requirements.txt
fastapi
pytest
pytest-asyncio
httpx
mysql-connector-python
jwt
uvicorn
multipart
passlib